Altered database structure
Changed the database to use InnoDB and UTF-8 instead of MyISAM and
latin1. Also rearranged the users table so the user_id field is first,
and expanded the password field to 255 characters as required by the
switch from md5 hashes to PHP's password_hash() function. SQL file also
updated as a direct copy of the database structure.